// src/app/(resources)/(generated)/types/Company.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Company = {
  /**
   * @type string, uuid
   */
  id: string;
  /**
   * @type string, uuid
   */
  userId: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @description Valor decimal como string
   * @type string
   */
  hourlyRate: string;
  /**
   * @type boolean
   */
  isActive: boolean;
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};